+++ src/Table/Table.directive.ts
@@ -5,7 +5,7 @@ import { DefaultTableState } from "./../TableState/DefaultTableState.class";
 import { SortOrder } from './../Sort/SortOrder.enum';
 import { ConfigurationProvider } from './../Configuration/ConfigurationProvider.class';
 import { IConfiguration } from './../Configuration/IConfiguration.interface';
-import { IDataPipeService } from './../Pipe/IDataPipeService.interface';
+import { IDataPipeService, IDataPipeFunction } from './../Pipe/IDataPipeService.interface';
 
 @Directive({
     selector: "[ptTable]"
@@ -30,11 +30,19 @@ export class TableDirective {
     displayArrayChange: EventEmitter<Array<any>> = new EventEmitter<Array<any>>();
 
     /*
+        Event for custom data-pipe implemented by component.
+        Only used if observer is present. Otherwise a IDataPipeService
+        is resolved from the injector.
+    */
+    @Output('ptDataPipe')
+    public dataPipe: EventEmitter<any> = new EventEmitter<any>();
+
+    /*
         two-way binding for ITableState
     */
-    @Input()
+    @Input('ptTableState')
     public tableState: ITableState;
-    @Output()
+    @Output('ptTableStateChange')
     tableStateChange: EventEmitter<ITableState> = new EventEmitter<ITableState>();
 
     /*
@@ -58,23 +66,36 @@ export class TableDirective {
     }
 
     ngOnInit() {
-        if (this.tableState){
+        if (this.tableState) {
             this.tableStateChange.emit(this.tableState);
         }
 
         this.getTableState();
+
+        if (this.dataPipe.observers.length > 0){
+            this.pipe();
+        }
     }
 
     ngOnChanges(changes: SimpleChanges): void {
-        if (changes['tableState'] && this.tableState){
+        var callPipe: boolean = false;
+
+        if (changes['tableState'] && this.tableState) {
             this.tableStateChange.emit(this.tableState);
         }
+        if (changes['dataPipe']) {
+            callPipe = true;
+        }
         if (changes['originalArray']) {
-            this.pipe();
+            callPipe = true;
         }
         if (changes['configurationOverride']) {
             this.dataPipeService = null;
             this.currentConfiguration = null;
+            callPipe = true;
+        }
+
+        if (callPipe) {
             this.pipe();
         }
     }
@@ -86,13 +107,6 @@ export class TableDirective {
             this.changeDetectorRef.detectChanges();
         }
         return this.tableState;
-    }
-
-    public doSearch(predicate: string, reverse: boolean) {
-        // update table state
-        // 
-
-        this.pipe();
     }
 
     public getConfiguration(): IConfiguration {
@@ -108,19 +122,31 @@ export class TableDirective {
         return this.currentConfiguration;
     }
 
+    public updateDisplayArray(results: Array<any>, totalItemCount: number): void {
+        this.tableState.pagination.totalItemCount = totalItemCount;
+
+        this.displayArray = results;
+        this.displayArrayChange.emit(this.displayArray);
+    }
+
     public pipe() {
         var state = this.getTableState();
         var config = this.getConfiguration();
+        var pipeResult: Promise<Array<any>>;
 
-        if (!this.dataPipeService) {
-            this.dataPipeService = this.injector.get(config.pipeServiceType);
+        if (this.dataPipe.observers.length > 0) {
+            this.dataPipe.emit([this, state, config]);
+            return;
         }
 
+        if (!this.dataPipeService)
+            this.dataPipeService = this.injector.get(config.pipeServiceType);
+
         this.dataPipeService.pipe(this.originalArray, state, config)
-        .then( (array : Array<any>) => {
-            this.displayArray = array;
-            this.displayArrayChange.emit(this.displayArray);
-            this.changeDetectorRef.detectChanges();
-        });
+            .then((array: Array<any>) => {
+                this.displayArray = array;
+                this.displayArrayChange.emit(this.displayArray);
+                this.changeDetectorRef.detectChanges();
+            });
     };
 }